requirements.txt
faster-whisper==1.0.3
pyannote.audio==3.1.1
torch>=2.2.0
torchaudio>=2.2.0
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.30.6
